crawled poll urls from realclearpolitics.com
I am including the simple crawlers and the result for this task in this commition. 
There are totall 6 files, 2 crawlers, 4 text files:
- rcp_poll_url_crawler_national.py  
    • crawler for national poll
- rcp_poll_url_crawler_state.py 
    • crawler for state poll
- rcp_poll_url_national.txt   
    • distinguished url list for national polls
- rcp_poll_url_state.txt    
    • distinguished url list for state polls
- crp_poll_domain_national.txt
    • distinguished domain list for national polls
- crp_poll_domain_state.txt
    • distinguished domain list for state polls

I tried to include the data of poll in the file, but I couldn't find a way to scrap the infomation. So, there is not data in the list.